requirements.txt
torch

torchvision

pillow

scikit-learn

transformers

numpy

pandas

matplotlib

seaborn
